Stdlib. int ReadAndDecodeConsoleKeys(void)

+++ include/c8080/console.h
@@ -51,4 +51,4 @@ void __global MoveCursor(uint8_t x, uint8_t y) __link("c8080/console_h/movecurso
 void WriteConsole(const char *text) __link("c8080/console_h/writeconsole.c");
 
 /* Read console input and replace multi-byte key codes with a constant of type "KEY_" of fixed type uint16_t */
-uint16_t ReadAndDecodeConsoleKeys(void) __link("c8080/console_h/readanddecodeconsolekeys.c");
+int ReadAndDecodeConsoleKeys(void) __link("c8080/console_h/readanddecodeconsolekeys.c");
